+++ crates/unftp-auth-jsonfile/src/lib.rs
@@ -146,6 +146,7 @@
 
 use async_trait::async_trait;
 use bytes::Bytes;
+use flate2::read::GzDecoder;
 use ipnet::Ipv4Net;
 use iprange::IpRange;
 use libunftp::auth::{AuthenticationError, Authenticator, DefaultUser};
@@ -154,6 +155,7 @@ use ring::{
     pbkdf2::{verify, PBKDF2_HMAC_SHA256},
 };
 use serde::Deserialize;
+use std::io::prelude::*;
 use std::{collections::HashMap, convert::TryInto, fs, num::NonZeroU32, path::Path, time::Duration};
 use tokio::time::sleep;
 use valid::{constraint::Length, Validate};
@@ -210,7 +212,40 @@ struct UserCreds {
 impl JsonFileAuthenticator {
     /// Initialize a new [`JsonFileAuthenticator`] from file.
     pub fn from_file<P: AsRef<Path>>(filename: P) -> Result<Self, Box<dyn std::error::Error>> {
-        let json: String = fs::read_to_string(filename)?;
+        let mut f = fs::File::open(&filename)?;
+
+        // The credentials file can be plaintext, gzipped, or gzipped+base64-encoded
+        // The gzip-base64 format is useful for overcoming configmap size limits in Kubernetes
+        let mut magic: [u8; 4] = [0; 4];
+        let n = f.read(&mut magic[..])?;
+        let is_gz = n > 2 && magic[0] == 0x1F && magic[1] == 0x8B && magic[2] == 0x8;
+        // the 3 magic bytes translate to "H4sI" in base64
+        let is_base64gz = n > 2 && magic[0] == b'H' && magic[1] == b'4' && magic[2] == b's' && magic[3] == b'I';
+
+        f.rewind()?;
+        let json: String = match is_gz | is_base64gz {
+            true => {
+                let mut gzdata: Vec<u8> = Vec::new();
+                if is_base64gz {
+                    let mut b = Vec::new();
+                    f.read_to_end(&mut b)?;
+                    b.retain(|&x| x != b'\n' && x != b'\r');
+                    gzdata = base64::decode(b)?;
+                } else {
+                    f.read_to_end(&mut gzdata)?;
+                }
+                let mut d = GzDecoder::new(&gzdata[..]);
+                let mut s = String::new();
+                d.read_to_string(&mut s)?;
+                s
+            }
+            false => {
+                let mut s = String::new();
+                f.read_to_string(&mut s)?;
+                s
+            }
+        };
+
         Self::from_json(json)
     }